Refactor SubmitPostComment function
