+++ src/blog-comment-function/src/BlogComments/Functions/Validation/IPostExistenceChecker.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace BlogComments.Functions.Validation
+{
+    public interface IPostExistenceChecker
+    {
+        Task<string?> TryGetPostFileNameFromRepositoryAsync(string postName);
+    }
+}
