+++ src/blog-comment-function/src/BlogComments/ISystemClock.cs
@@ -1,6 +1,6 @@
-﻿using System;
+using System;
 
-namespace BlogComments.GitHub
+namespace BlogComments
 {
     public interface ISystemClock
     {
